Remove FunctionalBasis
