+++ lib/src/Base/Func/CMakeLists.txt
@@ -114,8 +114,6 @@ ot_add_source_file (ParametricEvaluation.cxx)
 ot_add_source_file (ParametricGradient.cxx)
 ot_add_source_file (ParametricHessian.cxx)
 ot_add_source_file (P1LagrangeEvaluation.cxx)
-ot_add_source_file (FunctionalBasisImplementation.cxx)
-ot_add_source_file (FunctionalBasis.cxx)
 ot_add_source_file (UniVariateFunctionFamily.cxx)
 ot_add_source_file (UniVariateFunctionFactory.cxx)
 ot_add_source_file (TensorizedUniVariateFunctionFactory.cxx)
@@ -235,8 +233,6 @@ ot_install_header_file (ParametricEvaluation.hxx)
 ot_install_header_file (ParametricGradient.hxx)
 ot_install_header_file (ParametricHessian.hxx)
 ot_install_header_file (P1LagrangeEvaluation.hxx)
-ot_install_header_file (FunctionalBasisImplementation.hxx)
-ot_install_header_file (FunctionalBasis.hxx)
 ot_install_header_file (UniVariateFunctionFamily.hxx)
 ot_install_header_file (UniVariateFunctionFactory.hxx)
 ot_install_header_file (TensorizedUniVariateFunctionFactory.hxx)
