+++ lib/src/Base/Func/openturns/TensorizedUniVariateFunctionFactory.hxx
@@ -21,7 +21,7 @@
 #ifndef OPENTURNS_TENSORIZEDUNIVARIATEFUNCTIONFACTORY_HXX
 #define OPENTURNS_TENSORIZEDUNIVARIATEFUNCTIONFACTORY_HXX
 
-#include "openturns/FunctionalBasisImplementation.hxx"
+#include "openturns/BasisImplementation.hxx"
 #include "openturns/EnumerateFunction.hxx"
 #include "openturns/Function.hxx"
 #include "openturns/PersistentCollection.hxx"
@@ -38,7 +38,7 @@ BEGIN_NAMESPACE_OPENTURNS
  */
 
 class OT_API TensorizedUniVariateFunctionFactory
-  : public FunctionalBasisImplementation
+  : public BasisImplementation
 {
   CLASSNAME
 public:
