+++ python/doc/user_manual/functions.rst
@@ -68,7 +68,6 @@ Collection of functions
     :template: class.rst_t
 
     Basis
-    FunctionalBasis
     OrthogonalBasis
     BasisSequence
