+++ python/src/BasisImplementation_doc.i.in
@@ -145,8 +145,7 @@ OT_Basis_getSubBasis_doc
 Returns
 -------
 isOrthogonal : bool
-    `True` if the basis is functional i.e. if its terms are a solution
-    to an equation (e.g. a basis made up of Legendre functions)."
+    `True` if the basis is functional i.e. infinite."
 %enddef
 %feature("docstring") OT::BasisImplementation::isFunctional
 OT_Basis_isFunctional_doc
